requirements.txt
discord.py>=2.3.2
anthropic>=0.40.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
